requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Pillow>=10.0.0
gunicorn==21.2.0
numpy==1.26.4
opencv-python-headless==4.8.1.78